+++ lib/main.dart
@@ -1,21 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:get/get.dart';
-import 'package:pizza_store_app/admin/item_admin/PageAddItem.dart';
-import 'package:pizza_store_app/admin/PageAdmin.dart';
 import 'package:pizza_store_app/controllers/controller_dashboard_manager.dart';
 import 'package:pizza_store_app/controllers/controller_home.dart';
-import 'package:pizza_store_app/controllers/controller_orders_manager.dart';
 import 'package:pizza_store_app/controllers/controller_user.dart';
-import 'package:pizza_store_app/layouts/MainLayout.dart';
-import 'package:pizza_store_app/layouts/ManagerLayout.dart';
-import 'package:pizza_store_app/pages/profile/PageChangeInfo.dart';
-import 'package:pizza_store_app/pages/home/PageItem.dart';
-import 'package:pizza_store_app/pages/shipper/PageOrderDetails.dart';
-import 'package:pizza_store_app/pages/order_manager/PageOrdersList.dart';
-import 'package:pizza_store_app/pages/shipper/PagePendingOrder.dart';
 import 'package:pizza_store_app/layouts/MainLayout.dart' show LocationBinding;
-
+import 'package:pizza_store_app/layouts/MainLayout.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 
 import 'helpers/other.helper.dart';
