+++ lib/pages/shipper/PageOrderDetails.dart
@@ -1,6 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import '../../controllers/controller_order_detail.dart';
+import 'package:pizza_store_app/enums/OrderStatus.dart';
+import 'package:pizza_store_app/helpers/other.helper.dart';
+
+import '../../controllers/controller_order_detail_shipper.dart';
 
 class PageOrderDetails extends StatelessWidget {
   final String orderId;
@@ -9,10 +12,10 @@ class PageOrderDetails extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return GetBuilder<OrderDetailsController>(
-      init: OrderDetailsController(orderId: orderId),
+    return GetBuilder<OrderDetailsShipperController>(
+      init: OrderDetailsShipperController(orderId: orderId),
       builder: (controller) {
-        if (controller.isLoadingCustomerOrder || controller.isLoadingUser) {
+        if (controller.isLoadingCustomerOrder) {
           return const Scaffold(
             body: Center(child: CircularProgressIndicator()),
           );
@@ -26,8 +29,10 @@ class PageOrderDetails extends StatelessWidget {
 
         return Scaffold(
           appBar: AppBar(
-            title: const Text("Chi tiết đơn hàng",
-                style: TextStyle(fontWeight: FontWeight.bold)),
+            title: const Text(
+              "Chi tiết đơn hàng",
+              style: TextStyle(fontWeight: FontWeight.bold),
+            ),
             centerTitle: true,
           ),
           body: SingleChildScrollView(
@@ -56,21 +61,12 @@ class PageOrderDetails extends StatelessWidget {
 }
 
 class _CustomerInfoCard extends StatelessWidget {
-  final OrderDetailsController controller;
+  final OrderDetailsShipperController controller;
 
   const _CustomerInfoCard({required this.controller});
 
   @override
   Widget build(BuildContext context) {
-    if (controller.isLoadingUser) {
-      return const Center(child: CircularProgressIndicator());
-    }
-
-    final user = controller.customer;
-    if (user == null) {
-      return const Center(child: Text("Không có thông tin khách hàng"));
-    }
-
     return Card(
       elevation: 2,
       shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
@@ -83,7 +79,7 @@ class _CustomerInfoCard extends StatelessWidget {
               crossAxisAlignment: CrossAxisAlignment.start,
               children: [
                 Text(
-                  user.userName,
+                  controller.customerOrder!.customer!.userName,
                   style: const TextStyle(
                     fontSize: 18,
                     fontWeight: FontWeight.bold,
@@ -95,7 +91,7 @@ class _CustomerInfoCard extends StatelessWidget {
                     const Icon(Icons.phone, size: 16),
                     const SizedBox(width: 8),
                     Text(
-                      user.phoneNumber,
+                      controller.customerOrder!.customer!.phoneNumber,
                       style: const TextStyle(fontSize: 16),
                     ),
                   ],
@@ -104,7 +100,9 @@ class _CustomerInfoCard extends StatelessWidget {
             ),
             ElevatedButton.icon(
               onPressed: () {
-                controller.openPhoneDial(user.phoneNumber);
+                controller.openPhoneDial(
+                  controller.customerOrder!.customer!.phoneNumber,
+                );
               },
               icon: const Icon(Icons.call, size: 20),
               label: const Text("Gọi ngay"),
@@ -114,7 +112,10 @@ class _CustomerInfoCard extends StatelessWidget {
                 shape: RoundedRectangleBorder(
                   borderRadius: BorderRadius.circular(8),
                 ),
-                padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
+                padding: const EdgeInsets.symmetric(
+                  horizontal: 16,
+                  vertical: 12,
+                ),
               ),
             ),
           ],
@@ -125,7 +126,8 @@ class _CustomerInfoCard extends StatelessWidget {
 }
 
 class _OrderDetailsCard extends StatelessWidget {
-  final OrderDetailsController controller;
+  final OrderDetailsShipperController controller;
+
   _OrderDetailsCard({required this.controller});
 
   @override
@@ -138,10 +140,7 @@ class _OrderDetailsCard extends StatelessWidget {
       children: [
         const Text(
           "Chi tiết đơn hàng",
-          style: TextStyle(
-            fontSize: 18,
-            fontWeight: FontWeight.bold,
-          ),
+          style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
         ),
         const SizedBox(height: 12),
         Card(
@@ -165,21 +164,27 @@ class _OrderDetailsCard extends StatelessWidget {
                     ),
                     Expanded(
                       flex: 1,
-                      child: Text("SL",
-                          style: TextStyle(fontWeight: FontWeight.bold),
-                          textAlign: TextAlign.center),
+                      child: Text(
+                        "SL",
+                        style: TextStyle(fontWeight: FontWeight.bold),
+                        textAlign: TextAlign.center,
+                      ),
                     ),
                     Expanded(
                       flex: 2,
-                      child: Text("Đơn giá",
-                          style: TextStyle(fontWeight: FontWeight.bold),
-                          textAlign: TextAlign.end),
+                      child: Text(
+                        "Đơn giá",
+                        style: TextStyle(fontWeight: FontWeight.bold),
+                        textAlign: TextAlign.end,
+                      ),
                     ),
                     Expanded(
                       flex: 2,
-                      child: Text("Thành tiền",
-                          style: TextStyle(fontWeight: FontWeight.bold),
-                          textAlign: TextAlign.end),
+                      child: Text(
+                        "Thành tiền",
+                        style: TextStyle(fontWeight: FontWeight.bold),
+                        textAlign: TextAlign.end,
+                      ),
                     ),
                   ],
                 ),
@@ -202,23 +207,30 @@ class _OrderDetailsCard extends StatelessWidget {
                         ),
                         Expanded(
                           flex: 1,
-                          child: Text(orderDetail.amount.toString(),
-                              textAlign: TextAlign.center,
-                              style: const TextStyle(fontSize: 15)),
+                          child: Text(
+                            orderDetail.amount.toString(),
+                            textAlign: TextAlign.center,
+                            style: const TextStyle(fontSize: 15),
+                          ),
                         ),
                         Expanded(
                           flex: 2,
                           child: Text(
-                              "${orderDetail.item.price.toStringAsFixed(0)}đ",
-                              textAlign: TextAlign.end,
-                              style: const TextStyle(fontSize: 15)),
+                            formatMoney(money: orderDetail.item.price),
+                            textAlign: TextAlign.end,
+                            style: const TextStyle(fontSize: 15),
+                          ),
                         ),
                         Expanded(
                           flex: 2,
                           child: Text(
-                              "${(orderDetail.amount * orderDetail.item.price).toStringAsFixed(0)}đ",
-                              textAlign: TextAlign.end,
-                              style: const TextStyle(fontSize: 15)),
+                            formatMoney(
+                              money:
+                                  orderDetail.amount * orderDetail.item.price,
+                            ),
+                            textAlign: TextAlign.end,
+                            style: const TextStyle(fontSize: 15),
+                          ),
                         ),
                       ],
                     );
@@ -232,35 +244,45 @@ class _OrderDetailsCard extends StatelessWidget {
                     Row(
                       mainAxisAlignment: MainAxisAlignment.spaceBetween,
                       children: [
-                        const Text("Tạm tính:",
-                            style: TextStyle(fontSize: 15)),
+                        const Text("Tạm tính:", style: TextStyle(fontSize: 15)),
                         Text(
-                            "${controller.subTotal.toStringAsFixed(0)}đ",
-                            style: const TextStyle(fontSize: 15)),
+                          formatMoney(money: controller.subTotal),
+                          style: const TextStyle(fontSize: 15),
+                        ),
                       ],
                     ),
                     const SizedBox(height: 8),
                     Row(
                       mainAxisAlignment: MainAxisAlignment.spaceBetween,
                       children: [
-                        const Text("Phí vận chuyển:",
-                            style: TextStyle(fontSize: 15)),
+                        const Text(
+                          "Phí vận chuyển:",
+                          style: TextStyle(fontSize: 15),
+                        ),
                         Text(
-                            "${order.shippingFee?.toStringAsFixed(0) ?? "0"}đ",
-                            style: const TextStyle(fontSize: 15)),
+                          formatMoney(money: order.shippingFee ?? 0),
+                          style: const TextStyle(fontSize: 15),
+                        ),
                       ],
                     ),
                     const SizedBox(height: 8),
                     Row(
                       mainAxisAlignment: MainAxisAlignment.spaceBetween,
                       children: [
-                        const Text("Tổng cộng:",
-                            style: TextStyle(
-                                fontSize: 16, fontWeight: FontWeight.bold)),
+                        const Text(
+                          "Tổng cộng:",
+                          style: TextStyle(
+                            fontSize: 16,
+                            fontWeight: FontWeight.bold,
+                          ),
+                        ),
                         Text(
-                            "${controller.grandTotal.toStringAsFixed(0)}đ",
-                            style: const TextStyle(
-                                fontSize: 16, fontWeight: FontWeight.bold)),
+                          formatMoney(money: controller.grandTotal),
+                          style: const TextStyle(
+                            fontSize: 16,
+                            fontWeight: FontWeight.bold,
+                          ),
+                        ),
                       ],
                     ),
                   ],
@@ -275,7 +297,7 @@ class _OrderDetailsCard extends StatelessWidget {
 }
 
 class _DeliveryInfoCard extends StatelessWidget {
-  final OrderDetailsController controller;
+  final OrderDetailsShipperController controller;
 
   const _DeliveryInfoCard({required this.controller});
 
@@ -288,10 +310,7 @@ class _DeliveryInfoCard extends StatelessWidget {
       children: [
         const Text(
           "Thông tin giao hàng",
-          style: TextStyle(
-            fontSize: 18,
-            fontWeight: FontWeight.bold,
-          ),
+          style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
         ),
         const SizedBox(height: 12),
         Card(
@@ -306,7 +325,8 @@ class _DeliveryInfoCard extends StatelessWidget {
                 const Icon(Icons.location_on, color: Colors.red),
                 const SizedBox(width: 12),
                 Expanded(
-                  child: Text("${order.shippingAddress ?? "Không có địa chỉ "} ",
+                  child: Text(
+                    "${order.shippingAddress ?? "Không có địa chỉ "} ",
                     style: const TextStyle(fontSize: 16),
                   ),
                 ),
@@ -320,114 +340,83 @@ class _DeliveryInfoCard extends StatelessWidget {
 }
 
 class _ActionButtons extends StatelessWidget {
-  final OrderDetailsController controller;
+  final OrderDetailsShipperController controller;
 
   const _ActionButtons({required this.controller});
 
   @override
   Widget build(BuildContext context) {
-    return Obx(() {
-      final order = controller.customerOrder!;
-      final status = controller.orderStatus.value;
+    return GetBuilder<OrderDetailsShipperController>(
+      builder: (controller) {
+        final status = controller.customerOrder?.status;
 
-      if (status == 'finished') {
-        return const Center(
-          child: Text(
-            "Đơn hàng đã được giao",
-            style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
-          ),
+        if (status == OrderStatus.finished) {
+          return const Center(
+            child: Text(
+              "Đơn hàng đã được giao",
+              style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
+            ),
+          );
+        }
+
+        return Row(
+          children: [
+            const SizedBox(width: 16),
+            Expanded(
+              child:
+                  status == OrderStatus.shipping
+                      ? ElevatedButton(
+                        onPressed: () {
+                          controller.markOrderAsFinished();
+                        },
+                        style: ElevatedButton.styleFrom(
+                          backgroundColor: Colors.green,
+                          foregroundColor: Colors.white,
+                          padding: const EdgeInsets.symmetric(vertical: 16),
+                          shape: RoundedRectangleBorder(
+                            borderRadius: BorderRadius.circular(12),
+                          ),
+                        ),
+                        child: const Text("Hoàn tất đơn hàng"),
+                      )
+                      : ElevatedButton(
+                        onPressed: () {
+                          Get.dialog(
+                            AlertDialog(
+                              title: const Text("Xác nhận đơn hàng"),
+                              content: const Text(
+                                "Bạn có chắc chắn muốn xác nhận đơn hàng này?",
+                              ),
+                              actions: [
+                                TextButton(
+                                  onPressed: () => Get.back(),
+                                  child: const Text("Không"),
+                                ),
+                                TextButton(
+                                  onPressed: () {
+                                    controller.confirmOrder();
+                                    Get.back();
+                                  },
+                                  child: const Text("Có"),
+                                ),
+                              ],
+                            ),
+                          );
+                        },
+                        style: ElevatedButton.styleFrom(
+                          backgroundColor: Colors.blue,
+                          foregroundColor: Colors.white,
+                          padding: const EdgeInsets.symmetric(vertical: 16),
+                          shape: RoundedRectangleBorder(
+                            borderRadius: BorderRadius.circular(12),
+                          ),
+                        ),
+                        child: const Text("Xác nhận đơn hàng"),
+                      ),
+            ),
+          ],
         );
-      }
-
-      return Row(
-        children: [
-          Expanded(
-            child: ElevatedButton(
-              onPressed: () {
-                Get.dialog(
-                  AlertDialog(
-                    title: const Text("Xác nhận hủy đơn hàng"),
-                    content: const Text(
-                        "Bạn có chắc chắn muốn hủy đơn hàng này?"),
-                    actions: [
-                      TextButton(
-                        onPressed: () => Get.back(),
-                        child: const Text("Không"),
-                      ),
-                      TextButton(
-                        onPressed: () {
-                          Get.back();
-                        },
-                        child: const Text("Có"),
-                      ),
-                    ],
-                  ),
-                );
-              },
-              style: ElevatedButton.styleFrom(
-                backgroundColor: Colors.red,
-                foregroundColor: Colors.white,
-                padding: const EdgeInsets.symmetric(vertical: 16),
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(12),
-                ),
-              ),
-              child: const Text("Hủy đơn hàng"),
-            ),
-          ),
-          const SizedBox(width: 16),
-          Expanded(
-            child: status == 'shipping'
-                ? ElevatedButton(
-              onPressed: () {
-                controller.markOrderAsFinished();
-              },
-              style: ElevatedButton.styleFrom(
-                backgroundColor: Colors.green,
-                foregroundColor: Colors.white,
-                padding: const EdgeInsets.symmetric(vertical: 16),
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(12),
-                ),
-              ),
-              child: const Text("Hoàn tất đơn hàng"),
-            )
-                : ElevatedButton(
-              onPressed: () {
-                Get.dialog(
-                  AlertDialog(
-                    title: const Text("Xác nhận đơn hàng"),
-                    content: const Text(
-                        "Bạn có chắc chắn muốn xác nhận đơn hàng này?"),
-                    actions: [
-                      TextButton(
-                        onPressed: () => Get.back(),
-                        child: const Text("Không"),
-                      ),
-                      TextButton(
-                        onPressed: () {
-                          controller.confirmOrder();
-                          Get.back();
-                        },
-                        child: const Text("Có"),
-                      ),
-                    ],
-                  ),
-                );
-              },
-              style: ElevatedButton.styleFrom(
-                backgroundColor: Colors.blue,
-                foregroundColor: Colors.white,
-                padding: const EdgeInsets.symmetric(vertical: 16),
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(12),
-                ),
-              ),
-              child: const Text("Xác nhận đơn hàng"),
-            ),
-          ),
-        ],
-      );
-    });
+      },
+    );
   }
 }
